Preprocess the data as necessary before running SLISE

Preprocess the data as necessary before running SLISE

Usage

slise.preprocess(
  X,
  Y,
  epsilon,
  x = NULL,
  y = NULL,
  lambda1 = 0,
  lambda2 = 0,
  weight = NULL,
  intercept = FALSE,
  normalise = FALSE,
  logit = FALSE
)

Arguments

X

Matrix of independent variables

Y

Vector of the target variable

epsilon

Error tolerance

x

The sample to be explained (or index if y is null)

y

The prediction to be explained (default: NULL)

lambda1

L1 regularisation coefficient (default: 0)

lambda2

L2 regularisation coefficient (default: 0)

weight

Optional weight vector (default: NULL)

intercept

Should an intercept be added (default: TRUE)

normalise

Preprocess X and Y by scaling, note that epsilon is not scaled (default: FALSE)

logit

Logit transform Y from probabilities to real values (default: FALSE)

Value

list(X_orig, Y_orig, X, Y, x_orig, y_orig, x, y)